backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
alembic
python-dotenv
bcrypt
PyJWT 
# Optional dev/test
pytest
httpx
